Adapted LabelSubstitutionFF so that no Match features (NOMATCH) is added when a label subsitutes for the glue rule nonterminal [X].
A subsitution for the glue rule nonterminal  is not informative as a matching
or nonmatching substitution since  glue grammars for fuzzy_matching
systems only have X so subderivation top nodes will always have
to subsitute for it. Leaving out then this matching feature
for this case, were it is clearly not informative, adds to the information
the feature can give as signaling actual matching substitutions
(which we would intuitively belief to be better) and non-matching
substitutions (which we would intuitively expect to be less preferable).

	modified:   src/joshua/decoder/ff/LabelSubstitutionFF.java