Changed run-mira.pl so that the "continue" mode works properly to continue from the last completed + 1 iteration.
To do so, adapted it to read the initial weights from the last finished + 1 th joshua
config, rather than retrieving them in a convoluted and buggy way from the mert output file.
The latter leads to errors, as it is not certain that all feature weights will be outputted
in the n-best lists of the translations, as some (sparse) features might simply not fire.
This then leads to errors, when trying to use the "continue" flag with sparse features.
This fix simplifies things, and just uses the weights from the next iteration (which has not been
completed) as the starting point. This was probably also meant in the original code, but
it did not work out in that way, because the fact that the list of mert.out features might
not be complete was ignored

	modified:   scripts/training/mira/run-mira.pl