Added a new package featureScorePrediction, with new classes for the computation of predicted values for various types of label substitution features.
The plan is that we will predict the highest possible (partial)
score for label substitution features, given the feature weights
as presently used by the decoder. This optimistic prediction
will then help to quicly eliminate any collection of labeled rules.

This will be done by checking if the score for the first labeled variant, minus the
partial label substitution score and plus
the prediction for the highest possible label substitution score
is lower than the score for the candidate at position POP_LIMIT
in the queue. If so, non of the labeled variants will ever be
selected during cube pruning, so they can be discarded directly
without any loss, saving out a lot of computation.

	modified:   src/joshua/decoder/ff/LabelSubstitutionFF.java
	modified:   src/joshua/decoder/ff/LabelSubstitutionFFValuePrediction.java
	new file:   src/joshua/decoder/ff/featureScorePrediction/BasicLabelSubstitutionFeatureScorePredictor.java
	new file:   src/joshua/decoder/ff/featureScorePrediction/LabelMatchingFeatureScorePredictor.java
	new file:   src/joshua/decoder/ff/featureScorePrediction/SparseLabelSubstitutionFeatureScorePredictor.java