Further extended the implementation of label substitution feature score prediction. The main idea is that there is a hierarchy of score prediction classes, where a class higher in the hierarchy first calculates its own score and then recursively delegates to the next class lower in the hierarchy to get a complete score.
The main high level functions are  predictMaximumLabelSubstitutionFeatureScoreUnaryRule()
and predictMaxiLabelSubstitutionFeatureScoreBinaryRule(). Classes implement these functions
from the interface FeatureScorePredictor to be able to predict the best possible score
for unary and binary scores respectively.

The next steps in the implementation are to create and store a FeatureScorePredictor of the correct type
in an accesible place, so that the chart decoder has access to it.
Then the next step is to use the predictions to reject whole sets of candidates during cube-pruning,
candidates that are the same except for their labels, and can be predicted all to fall below the
top POP_LIMIT best candidates.

	modified:   src/joshua/decoder/ff/LabelSubstitutionFF.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/BasicLabelSubstitutionFeatureScorePredictor.java
	new file:   src/joshua/decoder/ff/featureScorePrediction/DoubleLabeledRulesLSFScorePredictor.java
	new file:   src/joshua/decoder/ff/featureScorePrediction/FeatureScorePredictor.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/LabelMatchingFeatureScorePredictor.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/SparseLabelSubstitutionFeatureScorePredictor.java
	new file:   src/joshua/decoder/ff/featureScorePrediction/SparseSubstitutionDescription.java