Further implemented the feature score prediction for label substitution features. Added a FeaureScorePredictor field to the decoder class, which allows it to be computed only once and be available everywhere in principle. Fixed some minor bugs.
Added a class LabelSubstitutionFeatureScorePredictionMaxScoresOnly.java
which is used to store the actual max scores computed by the actual
feature score predictors, so that it is not nescessary to keep big
tables in memory during decoding unnescessarily.

In this, we exploit the fact that the maximum binary and unary label
substitution score needs to be computed only once for every tuning
iteration based on the feature weights. After that all the created
temporary feature predictor classes can be discarded, and only the
two maximum scores need to be kept.

	modified:   src/joshua/decoder/Decoder.java
	modified:   src/joshua/decoder/ff/LabelSubstitutionFF.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/DoubleLabeledRulesLSFScorePredictor.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/LabelMatchingFeatureScorePredictor.java
	new file:   src/joshua/decoder/ff/featureScorePrediction/LabelSubstitutionFFAndFFPredictionCreater.java
	new file:   src/joshua/decoder/ff/featureScorePrediction/LabelSubstitutionFeatureScorePredictionMaxScoresOnly.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/SparseLabelSubstitutionFeatureScorePredictor.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/SparseSubstitutionDescription.java