Fixed the run-mira.pl script, with respect to the decoder command: Joshua only writes to standard output, hence to create the proper output for the tuner the output of the decoder needs to be piped to $JOSHUA/scripts/training/mira/feature_label_munger.pl to bring it into the right format.
It might be possible to
do this step inside a script, so that the run_mira.pl
script doesn't have to be changed. This script would then itself
have to retrieve the n-best list file name  and use it to to
do the munging and output of the translations to the standard output.
This might be clearner and could be changed on next commmits.

Note also that  src/joshua/decoder/Decoder.java
was adapted to convert also OOVPenalty, namely to OOV_Penalty.
Things did not work properly without this change.

  private String mosesize(String feature) {
    if (joshuaConfiguration.moses) {
      if (feature.startsWith("tm_") || feature.startsWith("lm_")){
        return feature.replace("_", "-");
      }
      if(feature.equals("OOVPenalty")){   <= NEW
        return "OOV_Penalty";
      }
    }

    return feature;
  }

Also a change has been made in the run_mira.pl script with respect to
checking the function sanity_check_order_of_lambdas:

sub sanity_check_order_of_lambdas {
  my $featlist = shift;
  my $filename_or_stream = shift;

  my @expected_lambdas = @{$featlist->{"names"}};
  my @got = get_order_of_scores_from_nbestlist($filename_or_stream);
  push @got, "OOV_Penalty";  ### Gideon add OOV_Penalty which is recognized as sparse feature and not returned by @got but included in @expected_lambdas   <======
  die "Mismatched lambdas. Decoder returned @got, we expected @expected_lambdas"
    if "@got" ne "@expected_lambdas";
}

because the function  get_order_of_scores_from_nbestlist ignores sparese features, it misses OOV_Penalty
which caused a mismatch. We added OOV_Penalty to the list @got to fix this.
This is a bit of a dirty fix, but until there is a better solution, this works at least.

	modified:   scripts/training/mira/run-mira.pl
	modified:   src/joshua/decoder/Decoder.java