Fixed the run_mira.pl script, so that it takes care of adding a flag "-moses true" to the Joshua decoder call.
This flag makes the decoder run in moses compatibility mode,
meaning that the feature names are adopted back and forth
when passing to the tuner in order to achieve compatibility
with the moses tuning script (run_mira.pl) which has
particular conventions for feature weight names.

It is better to put the flag in the run_mira.pl script than in the
configuration file, because when running the decoder "normally"
we don't want the flag, instead we just want it to run in normal
mode producing nbest lists in normal mode.

Note: one final possiblilty is not to adapt the run-mira.pl script
for this matter, but rather to add it as part of a specially designed
script for running Joshua for moses tuning.
This script could then add the parameter in the call, and also
use the feature munger and "tee" command in combination with
bash pipes to produce the NBest list in moses format as well
as the normal output (list of translations). Thus, we could
migrate the adaptations made to run-mira.pl instead to an
external script.
While this has the advantage of not having to adapt
run-mira.pl, we still would need to retain the change with respect
to adding the OOV_Penalty manually to the list of obtained features
from the N-best lists still.
Leave these changes for now and implement maybe after having discussed
with Matt Post.

$RUN_JOSHUA_IN_MOSES_COMPATIBLE_MODE_FLAG