Improved the decoder to compute the label substitution feature score predictor, and improved the label substitution feature score prediction classes.
Made a fix to scripts/training/mira/run-mira.pl:
this script was not using the updated joshua config produced during tuning iterations,
but rather the static configuration given at the beginning for each calll to the
decoder. This is problematic, since Joshua ignores the sparse weights file, and
therefore has no way to benefit from the updated weights for the sparse features
as produced by the tuner and stored in run*.joshua.config.

Note that an older version of run_mira.pl would simply overwrite the initial
config file, with the updated version. This is another way to achieve more
or less the same behavior. Here instead, we have adapted the decoder command
to use run*.joshua.config (i.e. run1.joshua.config, run2.joshua.config, ...).

	modified:   scripts/training/mira/run-mira.pl
	modified:   src/joshua/decoder/Decoder.java
	modified:   src/joshua/decoder/ff/LabelSubstitutionFF.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/LabelSubstitutionFFAndFFPredictionCreater.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/SparseLabelSubstitutionFeatureScorePredictor.java
	modified:   src/joshua/decoder/ff/featureScorePrediction/SparseSubstitutionDescription.java