Further improved and debugged the new implementation of fuzzy matching.
The first creation of CubePruneState objects was not yet correct
in the method addNewCandidate in Chart.java
What was not correct is that the initialization of the tailnodes
was not yet considering invalid HGNodes not corresponding to
one of the permissible nonterminal labels.
The updated method
  private  List<HGNode> getCurrentTailNodesForDotNode(T dotNode,
      int[] ranks, Chart<?, ?> chart)
takes care of these problems, using the new classes
ValidAntNodeComputer.java and NextAntNodesPreparer.java
to determine the first valid antnodes.
It also updates the ranks that are passed to the input,
in the case one or more of the first valide HGNodes is not
at the first index (0) of the corresponding Cell's HGNode list.

The new classes NextAntNodesPreparer.java
and ValidAntNodeComputer.java are motivated by the need to
compute the same things in both the method "addNewCandidate"
and "applyCubePruning" in Chart.java.
These classes allow for a more or less clean generalization of
computing the next state for cube pruning in case fuzzy matching
is done and we are dealing with DotNodeMultiLabel instead of
simple DotNode instances.

Also cleaned up the DotNodeBase, DotNode and DotNodeMultiLabel
classes, removing the duplicated lists for antnodes and replaced
with a list of the generic objects
      protected List<T> antSuperNodes = null;
in DotNodeBase.

Note that all joshua repository unit tests were ran and everything
except the two server-based tests (which did not pass before either)
passes.

	modified:   src/joshua/decoder/chart_parser/Chart.java
	modified:   src/joshua/decoder/chart_parser/CubePruneState.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateBase.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateFuzzyMatching.java
	modified:   src/joshua/decoder/chart_parser/DotChart.java
	new file:   src/joshua/decoder/chart_parser/NextAntNodesPreparer.java
	new file:   src/joshua/decoder/chart_parser/ValidAntNodeComputer.java