Made considerable changes to the cube pruning implementation for fuzzy matching. Impelemented a reorganization of the search space initialization and navigation by cube pruning.
Instead of completely disregarding the labels in the cube pruning state,
we create special states specialized for matching substitution.
When we initialize the cube pruning state for a rule, rather than just
creating one state which disregards the labels,
we also create additional states that fix one or both gap labels to
be the matching gap labels.
Furthermore, we adapt states working with
 non-matching gap labels, so that they skip over
gap labels that happen to match coincidentally.

The idea behind these changes, is that we want to give the search
a better chance to explore matching substitutions, which intuitively
should be a first default which can then be given up if it turns out
not to be good (i.e. the labels are bad).
Without this intervention it seems that the optimized cube pruning
no longer has enough chance to encounter matching substitutions.
Since it no longer unpacks all label substitution combinations for a rule, evaluating
each of them and creating a separate state for them during initialization,
and it has no kind of future cost it can use to predict the benefit of matching
substitution, it can effectively not learn to deliberately try matching
substitutions. Once these substitutions are evaluated, it can ofcourse determine
that they are good, but the evaluation itself may not happen sufficiently.

The hope is that these changes will considerably improve the search space explored,
without leading to too big increases in computational complexity
(in theory initialization before cube pruning could take 4 times longer at most,
but on the total running time we expect the effect to be much smaller).

	modified:   src/joshua/decoder/chart_parser/Chart.java
	modified:   src/joshua/decoder/chart_parser/CubePruneState.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateBase.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateFuzzyMatching.java
	modified:   src/joshua/decoder/chart_parser/NextAntNodesPreparer.java
	modified:   src/joshua/decoder/chart_parser/ValidAntNodeComputer.java