Fixed a bug in  src/joshua/decoder/chart_parser/ValidAntNodeComputer.java in the method    public static List<Integer> getAcceptableLabelIndicesNonterminalNotMatchingRuleLabel .
The method was directly comparing raw keys to check if nonterminals were matching,
however one of these keys is always negative and the other always positive, hence
this is incorrect.

Fixed this by using the method
 superNodeMatchesRuleNonterminal(SuperNode superNode,
      Rule rule, int nonterminalIndex)
from CubePruneStateFuzzyMatching, which uses the Vocabulary.word method
to lookup the actual labels, and perform a correct comparison of the labels
(effectively ignoring the sign of the keys).

The bug effectively caused the creation of a redundant cube pruning state space,
because the matching substitutions were both in the matching and mismatching
cube pruning states. It is not clear whether this bug caused serious problems
apart from clear inefficiency and a redundant search space.

	modified:   src/joshua/decoder/chart_parser/CubePruneStateBase.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateFuzzyMatching.java
	modified:   src/joshua/decoder/chart_parser/ValidAntNodeComputer.java