Fixed a bug in the CubePruneStateBase class. This class should now also consider the lists of acceptable nonterminal labels for the nonterminals in the equals and hashCode methods. The reason is that for specializations of the class using DotNodeMultiLabel, there can be different instances of the class, that share the same DotNodeMultiLabel, but that are to be considered distinct, because each of them is concerned with another subset of acceptable labels. The current example of this case is for the states that specialize in matching vs non-matching substitutions.
Not involving the acceptable nonterminal labels in the equals and
hashCode method consequently caused a bug, whereby in the cubepruning
code, some cubeprunestates are skipped, because the algorithm
incorrectly concludes they have already been added.
This should fix this, and hopefully will lead again to more matching
substitutions and better results
(currently the number of matching substitutions is surprisingly low,
while language model weights are high - the hope is that this bugfix
will change this).

	modified:   src/joshua/decoder/chart_parser/CubePruneStateBase.java