Changed the return value of the method getAcceptableLabelIndicesNonterminal() in ValidAntNodeComputer from List to set. This is nescessary to get a good performance when checking whether a nonterminal label is accetable or not, which needs to be done with every extension of a cube pruning state. Wiht a list this performancs is linear in the number of elements, with a (hash) set it is constant time.
-  public abstract List<Integer> getAcceptableLabelIndicesNonterminal();
+  public abstract Set<Integer> getAcceptableLabelIndicesNonterminal();

Hopefully this will solve most of the problems, were running time still
increases considerably with more labels in fuzzy matching decoding.
There is also still a call in the DotChart in the method extendDotItemsWithProvedItems,
which takes time linear in the number of labels. However, this is only called once
at the beginning of the method, so maybe it is still OK.

Another approach is to work with lists of unacceptable labels instead of acceptable
labels for fuzzy matching, i.e. keep the list of labels that need to be strictly matched.
This requires more drastic changes to the code however, so will be implemented as the
next step if needed.

	modified:   src/joshua/decoder/chart_parser/Chart.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateBase.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateFuzzyMatching.java
	modified:   src/joshua/decoder/chart_parser/ValidAntNodeComputer.java