Optimized the fuzzy matching performance, by specifying a set of indices that may not be matched as opposed to describing what may be matched.
It is still nescessary to separate NonOOV and Goal nonterminals
from the rest in DotChart, but in the cube pruning state
it is no longer nescessary to create a big set of acceptable
indices, so this should give a good improvement to performance.
Completely avoiding the separation between OOV and Goal symbols
and the rest in dotchart seems tricky: the problem is that the
two need to be separated, as the former needs strict matching,
and the latter needs soft matching.
It is not completely clear if it is nescessary to completely
avoid the separation (which takes time linear in the number of
labels, for every span in the dot chart) is nescessary.
Possibly also a gain can be made by hashing some values that
are combined multiple times in the dotchart.

	modified:   src/joshua/decoder/chart_parser/Chart.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateBase.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateFuzzyMatching.java
	modified:   src/joshua/decoder/chart_parser/DotChart.java
	modified:   src/joshua/decoder/chart_parser/DotNodeTypeCreater.java
	modified:   src/joshua/decoder/chart_parser/NonterminalMatcher.java
	modified:   src/joshua/decoder/chart_parser/ValidAntNodeComputer.java