Implemented and tested a variant of fuzzy matching without use of the dot chart (extending the existing code for decoding without dot chart, based on "A CYK+ Variant for SCFG Decoding Without A Dot Chart" (Sennrich, SSST    * 2014). ) This fuzzy matching implementation should be more efficient, and scale better to larger label sets.
This code completely avoids the linear time separation of OOV and Goal and non-OOV and
only collects the first non-OOV and non-Goal SuperNode, which does not add complexity.
The OOV and Goal symbol can be efficiently found in O(1) from the keyset of a cell,
which is backed by a HashMap (hence contains method has O(1) complexity for the keyset).
This is important, because this operation takes place deep inside the decoder loops, so
linear time operations are costly and best to avoid when possible.

	modified:   src/joshua/decoder/chart_parser/Chart.java
	modified:   src/joshua/decoder/chart_parser/CubePruneState.java
	modified:   src/joshua/decoder/chart_parser/CubePruneStateFuzzyMatching.java
	modified:   src/joshua/decoder/chart_parser/DotChart.java
	modified:   src/joshua/decoder/chart_parser/NonterminalMatcher.java
	modified:   src/joshua/decoder/chart_parser/ValidAntNodeComputer.java