Removed the check in the decoder that prevents num_translation_options > 0 (for specifying a finite number of translation options per rule) to be allowed.
With the new faster implementation of fuzzy matching, which removes labels
inside the Trie and only keeps them at the leaf level (for the rules
returned in the RuleCollection of leaf nodes, this may be useful.

In the faster fuzzy matching implementation, it is no longer true that
a rule is added in cube pruning inititialization for every occuring
labeling of the gap nonterminals that occurs in the grammar.
In the slower implementation, rules are separated by their labels,
and every differently labeled source variant that was able to match
is added during cube pruning initialization. This has the effect
of giving a high chance of producing at least one  matching substitution during
initialization, since effectively the rule labels are varied but the
labels for the best gap items remain constant during initialization.

The faster implementation seems to suffer a bit from having a lower guarantee
of properly exploring different (in particular matching?) label substitutions.
Possibly, by limiting the number of translation options per source side, we
can force more effort on a limited number of rules (with fixed labels)
and thus again give higher chance that both matching and non-matching substitutions
are explored for this limited number of rules
(an often used parameter value, also in moses is 20 options per rule).
Whether this approach works needs to be tested empirically.

	modified:   src/joshua/decoder/chart_parser/Chart.java