Added shuffling to the preparation of rules for evaluation and adding to the cube pruning queue. The motivation is the case of working with the Trie in which labels are not kept at first, but at second we do add different RuleCollections for alternative labelings of the same Hiero rule type. In this scenario, all different labeled versions of the same Hiero rule are evaluated in order, that is together.
This is potentially problematic, as it puts them all together in the queue,
if the evaluation score is insufficient to distinguish them from each
other or from other items.
By shuffling beforehand, we hope to increase the diversity of the queue
in cases where the item scores are not (yet) informative enough, which
may be especially true at the beginning of the tuning process.
An additional benefit is that this may indirectly counter overfitting
somewhat, because by avoiding that all labeled versions of a rule
are together and included/excluded  instead of mixed with other items
with same scores, it is plausible that we can reduce the opportunity
of overfitting with label substitution features.

So far, working with labels removed from the Trie, but recovering and
adding all different versions of Hiero rule types during cube pruning
initialization  has given worse results than working with the alternative
labeled versions inside the Trie, which yields an implicit form of
shuffling. By adding explicit shuffling, it is hoped that the performance
difference can be resolved, without having to go back to using
labeled rule versions inside the Trie, which is hugely inefficient
computationally when doing soft matching.
But whether the hypothesized problem is really the cause of the difference,
and whether this is thus resolved by shuffling, remains to be tested.

	modified:   src/joshua/decoder/chart_parser/Chart.java