Further implemented the restriction of the number of labeled variants per Language Model state.
Added production of statistics about the node addition behavior
and debugged the code.

Added a HGNode Signature map to RestrictLabeledVersionsLanguageModelStatePruning,
which is nescessary to see if nodes with a certain signature already exist.
It is not possible to reuse the HGNode signature map from Cell,
because the nodes that are present in the map in RestrictLabeledVersionsLanguageModelStatePruning
can change because of removing nodes.

Added some assertions to the new code to check that nodes are sorted in the right way, based on the highest log
probability.

	modified:   src/joshua/decoder/chart_parser/Cell.java
	modified:   src/joshua/decoder/chart_parser/Chart.java
	modified:   src/joshua/decoder/chart_parser/RestrictLabeledVersionsLanguageModelStatePruning.java
	modified:   src/joshua/decoder/chart_parser/RestrictedSizeBestHGNodeSet.java