Added cpu timing functionality to the decoder. To use the functionality, the config property
	log_cpu_computation_times = true
and
	cpu_computation_times_log_file_path = SOME_LOG_FILE_PATH

must be set in the joshua configuration that is passed to the decoder.

When we don't want to do the cpu timing, the bogus class ThreadCpuTimeKeeperBogus
is passed to the DecoderThreadRunner class (inside the Decoder.java file)
which does the actual decoding work that needs to be timed.
This way we can keep the same arguments in all involved methods, and don't
have to create spurious duplicate versions of methods and/or classes.

In the current implementation, the functionality is only usable for decoding
when the decoder is ran separately, e.g. for the test set, but not
for decoding as part of the tuning script.

To use the new code during tuning, we need to dynamically generating a log file name
for the i-th run in of the decoder, inside the tuning script.
This is somewhat more involved, it involves adapting the (mira) tuning script.
But we don't really need the cpu timing for the tuning
currently, doing the timing for the test set is sufficient.
We therefore leave this generalization (cpu timing during tuning as well) for future work.